0006: Content-Length の型を u64 に変更する
- HttpHead / Request / Response の content_length() を Option<u64> に変更する
- BodyKind::ContentLength と DecodePhase::BodyContentLength を u64 に変更する
- parse_content_length / resolve_body_headers を u64 に対応させる
- determine_body_kind / available_body_len / consume_body を u64 対応にする
- PBT / fuzz / skills ドキュメントを u64 に対応させる
- CHANGES.md に破壊的変更を記載する

+++ fuzz/fuzz_targets/fuzz_request_response_helpers.rs
@@ -45,8 +45,8 @@ fn header_count(headers: &[(String, String)], name: &str) -> usize {
         .count()
 }
 
-fn expected_content_length(headers: &[(String, String)]) -> Option<usize> {
-    header_value(headers, "Content-Length").and_then(|v| v.parse::<usize>().ok())
+fn expected_content_length(headers: &[(String, String)]) -> Option<u64> {
+    header_value(headers, "Content-Length").and_then(|v| v.parse::<u64>().ok())
 }
 
 fn expected_chunked(headers: &[(String, String)]) -> bool {
